Refactor geometry utilities into reusable module
