+++ appV5.py
@@ -71,6 +71,8 @@ import tkinter.font as tkfont
 import urllib.request
 from importlib import import_module
 from typing import Any, Dict, Iterable, List, Optional, Sequence, Tuple
+
+from cad_quoter import geometry
 from OCP.TopAbs   import TopAbs_EDGE, TopAbs_FACE
 from OCP.TopExp   import TopExp, TopExp_Explorer
 from OCP.TopTools import TopTools_IndexedDataMapOfShapeListOfShape
@@ -85,6 +87,32 @@ try:
     from geo_read_more import build_geo_from_dxf as build_geo_from_dxf_path
 except Exception:
     build_geo_from_dxf_path = None  # type: ignore[assignment]
+
+
+# Geometry helpers (re-exported for backward compatibility)
+load_model = geometry.load_model
+load_cad_any = geometry.load_cad_any
+read_cad_any = geometry.read_cad_any
+read_step_shape = geometry.read_step_shape
+read_step_or_iges_or_brep = geometry.read_step_or_iges_or_brep
+convert_dwg_to_dxf = geometry.convert_dwg_to_dxf
+enrich_geo_occ = geometry.enrich_geo_occ
+enrich_geo_stl = geometry.enrich_geo_stl
+safe_bbox = geometry.safe_bbox
+safe_bounding_box = geometry.safe_bounding_box
+iter_solids = geometry.iter_solids
+explode_compound = geometry.explode_compound
+parse_hole_table_lines = geometry.parse_hole_table_lines
+extract_text_lines_from_dxf = geometry.extract_text_lines_from_dxf
+upsert_var_row = geometry.upsert_var_row
+require_ezdxf = geometry.require_ezdxf
+get_dwg_converter_path = geometry.get_dwg_converter_path
+have_dwg_support = geometry.have_dwg_support
+get_import_diagnostics_text = geometry.get_import_diagnostics_text
+_HAS_TRIMESH = geometry.HAS_TRIMESH
+_HAS_EZDXF = geometry.HAS_EZDXF
+_HAS_ODAFC = geometry.HAS_ODAFC
+_EZDXF_VER = geometry.EZDXF_VERSION
 
 
 def _normalize_lookup_key(value: str) -> str:
@@ -2199,1376 +2227,6 @@ def find_default_qwen_model():
     choice = _pick_best_gguf([str(p) for p in ggufs])
     if choice: return choice
     return ""
-
-# Optional trimesh for STL
-try:
-    import trimesh  # type: ignore
-    _HAS_TRIMESH = True
-except Exception:
-    _HAS_TRIMESH = False
-
-_HAS_EZDXF  = False
-_HAS_ODAFC  = False      # ezdxf.addons.odafc (uses ODA File Converter)
-_EZDXF_VER  = "unknown"
-
-try:
-    import ezdxf
-    _EZDXF_VER = getattr(ezdxf, "__version__", "unknown")
-    _HAS_EZDXF = True
-except Exception:
-    ezdxf = None  # keep name defined
-
-# odafc is optional; don't fail if it's not present
-try:
-    if _HAS_EZDXF:
-        from ezdxf.addons import odafc  # type: ignore
-        _HAS_ODAFC = True
-except Exception:
-    _HAS_ODAFC = False
-
-# Hole chart helpers (optional)
-try:
-    from hole_table_parser import parse_hole_table_lines
-except Exception:
-    parse_hole_table_lines = None
-
-try:
-    from dxf_text_extract import extract_text_lines_from_dxf
-except Exception:
-    extract_text_lines_from_dxf = None
-
-# numpy is optional for a few small calcs; degrade gracefully if missing
-try:
-    import numpy as np  # type: ignore
-except Exception:
-    np = None  # type: ignore
-# ---------- OCC / OCP compatibility ----------
-STACK = None
-
-def _make_bnd_add_ocp():
-    # Try every known symbol name in OCP builds
-    candidates = []
-    try:
-        # module-level function: Add(...)
-        from OCP.BRepBndLib import Add as _add1
-        candidates.append(_add1)
-    except Exception:
-        pass
-    try:
-        # module-level function: Add_s(...)
-        from OCP.BRepBndLib import Add_s as _add1s
-        candidates.append(_add1s)
-    except Exception:
-        pass
-    try:
-        # module-level function: BRepBndLib_Add(...)
-        from OCP.BRepBndLib import BRepBndLib_Add as _add2
-        candidates.append(_add2)
-    except Exception:
-        pass
-    try:
-        # module-level function: brepbndlib_Add(...)
-        from OCP.BRepBndLib import brepbndlib_Add as _add3
-        candidates.append(_add3)
-    except Exception:
-        pass
-    try:
-        from OCP.BRepBndLib import BRepBndLib as _klass
-        for attr in ("Add", "Add_s", "AddClose_s", "AddOptimal_s", "AddOBB_s"):
-            fn = getattr(_klass, attr, None)
-            if fn:
-                candidates.append(fn)
-    except Exception:
-        pass
-
-    if not candidates:
-        raise ImportError("No BRepBndLib.Add symbol found in OCP")
-
-    # return a dispatcher that tries different call signatures
-    def _bnd_add(shape, box, use_triangulation=True):
-        last_err = None
-        for fn in candidates:
-            # Try 3-arg form
-            try:
-                return fn(shape, box, use_triangulation)
-            except TypeError as e:
-                last_err = e
-            # Try 2-arg form
-            try:
-                return fn(shape, box)
-            except TypeError as e:
-                last_err = e
-            except Exception as e:
-                last_err = e
-        # If we get here, none of the variants worked
-        raise last_err or RuntimeError("No callable BRepBndLib.Add variant succeeded")
-
-    return _bnd_add
-
-try:
-    # Prefer OCP (CadQuery/ocp bindings)
-    from OCP.STEPControl import STEPControl_Reader
-    from OCP.IFSelect import IFSelect_RetDone
-    from OCP.TopoDS import TopoDS_Compound, TopoDS_Shape
-    from OCP.BRep import BRep_Builder
-    from OCP.ShapeFix import ShapeFix_Shape
-    from OCP.BRepCheck import BRepCheck_Analyzer
-    from OCP.Bnd import Bnd_Box
-
-    try:
-        bnd_add = _make_bnd_add_ocp()
-        STACK = "ocp"
-    except Exception:
-        # Fallback: try dynamic dispatch at call time for OCP builds missing Add
-        def bnd_add(shape, box, use_triangulation=True):
-            candidates = [
-                ("OCP.BRepBndLib", "Add"),
-                ("OCP.BRepBndLib", "Add_s"),
-                ("OCP.BRepBndLib", "BRepBndLib_Add"),
-                ("OCP.BRepBndLib", "brepbndlib_Add"),
-                ("OCC.Core.BRepBndLib", "Add"),
-                ("OCC.Core.BRepBndLib", "Add_s"),
-                ("OCC.Core.BRepBndLib", "BRepBndLib_Add"),
-                ("OCC.Core.BRepBndLib", "brepbndlib_Add"),
-            ]
-            for mod_name, attr in candidates:
-                try:
-                    module = __import__(mod_name, fromlist=[attr])
-                    fn = getattr(module, attr)
-                    return fn(shape, box, use_triangulation)
-                except Exception:
-                    continue
-            raise RuntimeError("No BRepBndLib.Add available in this build")
-except Exception:
-    # Fallback to pythonocc-core
-    from OCC.Core.STEPControl import STEPControl_Reader
-    from OCC.Core.IFSelect import IFSelect_RetDone
-    from OCC.Core.TopoDS import TopoDS_Compound, TopoDS_Shape
-    from OCC.Core.BRep import BRep_Builder
-    from OCC.Core.ShapeFix import ShapeFix_Shape
-    from OCC.Core.BRepCheck import BRepCheck_Analyzer
-    from OCC.Core.Bnd import Bnd_Box
-    from OCC.Core.BRepBndLib import brepbndlib_Add as _brep_add
-
-    def bnd_add(shape, box, use_triangulation=True):
-        _brep_add(shape, box, use_triangulation)
-
-    STACK = "pythonocc"
-# ---------- end shim ----------
-# ----- one-backend imports -----
-try:
-    from OCP.BRep import BRep_Tool
-    from OCP.TopAbs import TopAbs_FACE
-    from OCP.TopExp import TopExp_Explorer
-    from OCP.TopLoc import TopLoc_Location
-    BACKEND = "ocp"
-except Exception:
-    from OCC.Core.BRep import BRep_Tool
-    from OCC.Core.TopoDS import topods_Edge, topods_Shell, topods_Solid
-    from OCC.Core.TopAbs import TopAbs_FACE
-    from OCC.Core.TopExp import TopExp_Explorer
-    from OCC.Core.TopLoc import TopLoc_Location
-    BACKEND = "pythonocc"
-
-def _typename(o):  # small helper
-    return getattr(o, "__class__", type(o)).__name__
-
-def as_face(obj):
-    """
-    Return a TopoDS_Face for the *current backend*.
-    - If already a Face, return it (don't re-cast).
-    - If a tuple (some APIs return (surf, loc) or similar), take first elem.
-    - If null/None, raise cleanly.
-    - Otherwise cast with FACE_OF.
-    """
-    if obj is None:
-        raise TypeError("Expected a shape, got None")
-
-    # unwrap tuples (defensive)
-    if isinstance(obj, tuple) and obj:
-        obj = obj[0]
-    return ensure_face(obj)
-
-def iter_faces(shape):
-    exp = TopExp_Explorer(shape, TopAbs_FACE)
-    while exp.More():
-        yield ensure_face(exp.Current())
-        exp.Next()
-
-def face_surface(face_like):
-    f = ensure_face(face_like)
-    if hasattr(BRep_Tool, "Surface_s"):
-        s = BRep_Tool.Surface_s(f)
-    else:
-        s = BRep_Tool.Surface(f)
-    loc = (BRep_Tool.Location(f) if hasattr(BRep_Tool, "Location")
-           else (f.Location() if hasattr(f, "Location") else None))
-    if isinstance(s, tuple):
-        s, loc2 = s
-        loc = loc or loc2
-    if hasattr(s, "Surface"):
-        s = s.Surface()  # unwrap handle
-    return s, loc
- 
- 
-    
-# ---------- OCCT compat (OCP or pythonocc-core) ----------
-# ---------- Robust casters that work on OCP and pythonocc ----------
-# Lock topods casters to the active backend
-if STACK == "ocp":
-    from OCP.TopoDS import TopoDS_Edge, TopoDS_Solid, TopoDS_Shell
-
-    def _TO_EDGE(s):
-        if type(s).__name__ in ("TopoDS_Edge", "Edge"):
-            return s
-        if hasattr(TopoDS, "Edge_s"):
-            return TopoDS.Edge_s(s)
-        try:
-            from OCP.TopoDS import topods as _topods
-            return _topods.Edge(s)
-        except Exception as e:
-            raise TypeError(f"Cannot cast to Edge from {type(s).__name__}") from e
-
-    def _TO_SOLID(s):
-        if type(s).__name__ in ("TopoDS_Solid", "Solid"):
-            return s
-        if hasattr(TopoDS, "Solid_s"):
-            return TopoDS.Solid_s(s)
-        from OCP.TopoDS import topods as _topods
-        return _topods.Solid(s)
-
-    def _TO_SHELL(s):
-        if type(s).__name__ in ("TopoDS_Shell", "Shell"):
-            return s
-        if hasattr(TopoDS, "Shell_s"):
-            return TopoDS.Shell_s(s)
-        from OCP.TopoDS import topods as _topods
-        return _topods.Shell(s)
-else:
-    # Resolve within OCC.Core only
-    try:
-        from OCC.Core.TopoDS import topods_Edge as _TO_EDGE
-    except Exception:
-        try:
-            from OCC.Core.TopoDS import Edge as _TO_EDGE
-        except Exception:
-            from OCC.Core.TopoDS import topods as _occ_topods
-            _TO_EDGE = getattr(_occ_topods, "Edge")
-    try:
-        from OCC.Core.TopoDS import topods_Solid as _TO_SOLID
-    except Exception:
-        from OCC.Core.TopoDS import topods as _occ_topods
-        _TO_SOLID = getattr(_occ_topods, "Solid")
-    try:
-        from OCC.Core.TopoDS import topods_Shell as _TO_SHELL
-    except Exception:
-        from OCC.Core.TopoDS import topods as _occ_topods
-        _TO_SHELL = getattr(_occ_topods, "Shell")
-
-# Type guards
-def _is_named(obj, names: tuple[str, ...]) -> bool:
-    try:
-        return type(obj).__name__ in names
-    except Exception:
-        return False
-
-def _unwrap_value(obj):
-    # Some containers return nodes with .Value()
-    return obj.Value() if hasattr(obj, "Value") and callable(obj.Value) else obj
-
-def ensure_shape(obj):
-    obj = _unwrap_value(obj)
-    if obj is None:
-        raise TypeError("Expected TopoDS_Shape, got None")
-    if hasattr(obj, "IsNull") and obj.IsNull():
-        raise TypeError("Expected non-null TopoDS_Shape")
-    return obj
-
-# Safe casters: no-ops if already cast; unwrap list nodes; check kind
-def to_edge_safe(obj):
-    obj = _unwrap_value(obj)
-    if _is_named(obj, ("TopoDS_Edge", "Edge")):
-        return obj
-    obj = ensure_shape(obj)
-    return _TO_EDGE(obj)
-
-# Choose stack
-try:
-    # OCP / CadQuery bindings
-    from OCP.BRepGProp import BRepGProp as _BRepGProp_mod
-    STACK_GPROP = "ocp"
-except Exception:
-    # pythonocc-core
-    try:
-        from OCC.Core.BRepGProp import BRepGProp as _BRepGProp_mod
-    except Exception:
-        from types import SimpleNamespace
-        from OCC.Core.BRepGProp import (
-            brepgprop_LinearProperties as _lp,
-            brepgprop_SurfaceProperties as _sp,
-            brepgprop_VolumeProperties as _vp,
-        )
-        _BRepGProp_mod = SimpleNamespace(
-            LinearProperties=_lp,
-            SurfaceProperties=_sp,
-            VolumeProperties=_vp,
-        )
-    def _to_edge_occ(s):
-        try:
-            from OCC.Core.TopoDS import topods_Edge as _fn
-        except Exception:
-            from OCC.Core.TopoDS import Edge as _fn
-        return _fn(s)
-
-    _TO_EDGE = _to_edge_occ
-    STACK_GPROP = "pythonocc"
-
-# Resolve topods casters across bindings
-
-
-# ---- modern wrappers (no deprecation warnings)
-# ---- modern wrappers (no deprecation warnings)
-def linear_properties(edge, gprops):
-    """Linear properties across OCP/pythonocc names."""
-    fn = getattr(_BRepGProp_mod, "LinearProperties", None)
-    if fn is None:
-        fn = getattr(_BRepGProp_mod, "LinearProperties_s", None)
-    if fn is None:
-        try:
-            from OCC.Core.BRepGProp import brepgprop_LinearProperties as _old  # type: ignore
-            return _old(edge, gprops)
-        except Exception as e:
-            raise
-    return fn(edge, gprops)
-
-def map_shapes_and_ancestors(root_shape, sub_enum, anc_enum):
-    """Return TopTools_IndexedDataMapOfShapeListOfShape for (sub → ancestors)."""
-    # Ensure we pass a *Shape*, not a Face
-    if root_shape is None:
-        raise TypeError("root_shape is None")
-    if not hasattr(root_shape, "IsNull") or root_shape.IsNull():
-        # If someone handed us a Face, try to grab its TShape parent; else fail.
-        # Safer: require a real TopoDS_Shape from STEP/IGES root.
-        pass
-
-    amap = TopTools_IndexedDataMapOfShapeListOfShape()
-    # static/instance variants across wheels
-    fn = getattr(TopExp, "MapShapesAndAncestors", None) or getattr(TopExp, "MapShapesAndAncestors_s", None)
-    if fn is None:
-        raise RuntimeError("TopExp.MapShapesAndAncestors not available in this OCP wheel")
-    fn(root_shape, sub_enum, anc_enum, amap)
-    return amap
-
-# modern topods casters: topods.Edge(shape) / topods.Face(shape)
-# ---- Robust topods casters that are no-ops for already-cast objects ----
-def _is_instance(obj, qualnames):
-    """True if obj.__class__.__name__ matches any in qualnames across OCP/OCC."""
-    try:
-        name = type(obj).__name__
-    except Exception:
-        return False
-    return name in qualnames  # e.g. ["TopoDS_Face", "Face"]
-
-def ensure_face(obj):
-    if obj is None:
-        raise TypeError("Expected a face, got None")
-    if isinstance(obj, TopoDS_Face) or type(obj).__name__ == "TopoDS_Face":
-        return obj
-    st = obj.ShapeType() if hasattr(obj, "ShapeType") else None
-    if st == TopAbs_FACE:
-        return FACE_OF(obj)
-    raise TypeError(f"Not a face: {type(obj).__name__}")
-def to_edge(s):
-    if _is_instance(s, ["TopoDS_Edge", "Edge"]):
-        return s
-    return _TO_EDGE(s)
-
-def to_solid(s):
-    if _is_instance(s, ["TopoDS_Solid", "Solid"]):
-        return s
-    return _TO_SOLID(s)
-
-def to_shell(s):
-    if _is_instance(s, ["TopoDS_Shell", "Shell"]):
-        return s
-    return _TO_SHELL(s)
-
-# Generic size helpers so we never call Extent() on lists/maps again
-def map_size(m):
-    for name in ("Size", "Extent", "Length"):
-        if hasattr(m, name):
-            return getattr(m, name)()
-    raise AttributeError(f"No size method on {type(m)}")
-
-def list_iter(lst):
-    """
-    Iterate TopTools_ListOfShape across bindings.
-    Use cbegin()/More()/Value()/Next() if available; otherwise fall back to python list conversion.
-    """
-    # OCP + pythonocc expose cbegin iterator on TopTools_ListOfShape
-    if hasattr(lst, "cbegin"):
-        it = lst.cbegin()
-        while it.More():
-            yield it.Value()
-            it.Next()
-    else:
-        # last-resort: try Python iteration (some wheels support it)
-        for v in list(lst):
-            yield v
-# ---------- end compat ----------
-
-
-# ---- tiny helpers you can use elsewhere --------------------------------------
-def require_ezdxf():
-    """Raise a clear error if ezdxf is missing."""
-    if not _HAS_EZDXF:
-        raise RuntimeError("ezdxf not installed. Install with pip/conda (package name: 'ezdxf').")
-    return ezdxf
-
-def get_dwg_converter_path() -> str:
-    """Resolve a DWG?DXF converter path (.bat/.cmd/.exe)."""
-    exe = os.environ.get("ODA_CONVERTER_EXE") or os.environ.get("DWG2DXF_EXE")
-    # Fall back to a local wrapper next to this script if it exists.
-    local = str(Path(__file__).with_name("dwg2dxf_wrapper.bat"))
-    if not exe and Path(local).exists():
-        exe = local
-    return exe or ""
-
-def have_dwg_support() -> bool:
-    """True if we can open DWG (either odafc or an external converter is available)."""
-    return _HAS_ODAFC or bool(get_dwg_converter_path())
-def get_import_diagnostics_text() -> str:
-    import sys, shutil, os
-    lines = []
-    lines.append(f"Python: {sys.executable}")
-    try:
-        import fitz  # PyMuPDF
-        lines.append("PyMuPDF: OK")
-    except Exception as e:
-        lines.append(f"PyMuPDF: MISSING ({e})")
-
-    try:
-        import ezdxf
-        lines.append(f"ezdxf: {getattr(ezdxf, '__version__', 'unknown')}")
-        try:
-            from ezdxf.addons import odafc
-            lines.append("ezdxf.addons.odafc: OK")
-        except Exception as e:
-            lines.append(f"ezdxf.addons.odafc: not available ({e})")
-    except Exception as e:
-        lines.append(f"ezdxf: MISSING ({e})")
-
-    oda = os.environ.get("ODA_CONVERTER_EXE") or "(not set)"
-    d2d = os.environ.get("DWG2DXF_EXE") or "(not set)"
-    lines.append(f"ODA_CONVERTER_EXE: {oda}")
-    lines.append(f"DWG2DXF_EXE: {d2d}")
-
-    # local wrapper presence
-    from pathlib import Path
-    wrapper = Path(__file__).with_name("dwg2dxf_wrapper.bat")
-    lines.append(f"Local wrapper present: {wrapper.exists()} ({wrapper})")
-    return "\n".join(lines)
-# Optional PDF stack
-try:
-    import fitz  # PyMuPDF
-    _HAS_PYMUPDF = True
-except Exception:
-    _HAS_PYMUPDF = False
-def upsert_var_row(df, item, value, dtype="number"):
-    """
-    Upsert one row by Item name (case-insensitive).
-    - Forces `item` and `value` to scalars.
-    - Works on the sanitized 3-column df (Item, Example..., Data Type...).
-    """
-    import numpy as np
-    import pandas as pd
-
-    # force scalars
-    if isinstance(item, pd.Series):
-        item = item.iloc[0]
-    item = str(item)
-
-    if isinstance(value, pd.Series):
-        value = value.iloc[0]
-    # try to make numeric if it looks numeric; otherwise keep as-is
-    try:
-        if value is None or (isinstance(value, float) and np.isnan(value)):
-            pass
-        else:
-            value = float(value)
-    except Exception:
-        # leave non-numerics (e.g., text) as-is
-        pass
-
-    # build a row with the exact df schema
-    cols = list(df.columns)
-    row = {c: "" for c in cols}
-    if "Item" in row:
-        row["Item"] = item
-    if "Example Values / Options" in row:
-        row["Example Values / Options"] = value
-    if "Data Type / Input Method" in row:
-        row["Data Type / Input Method"] = dtype
-
-    # case-insensitive exact match on Item
-    m = df["Item"].astype(str).str.casefold() == item.casefold()
-    if m.any():
-        df.loc[m, cols] = [row[c] for c in cols]
-        return df
-
-    # append
-    new_row = pd.DataFrame([[row[c] for c in cols]], columns=cols)
-    return pd.concat([df, new_row], ignore_index=True)
-
-DIM_RE = re.compile(r"(?:ï¿½|DIAM|DIA)\s*([0-9.+-]+)|R\s*([0-9.+-]+)|([0-9.+-]+)\s*[xX]\s*([0-9.+-]+)")
-
-def load_drawing(path: Path) -> Drawing:
-    require_ezdxf()
-    if path.suffix.lower() == ".dwg":
-        # Prefer explicit converter/wrapper if configured (works even if ODA isnï¿½t on PATH)
-        exe = get_dwg_converter_path()
-        if exe:
-            dxf_path = convert_dwg_to_dxf(str(path))
-            return ezdxf.readfile(dxf_path)
-        # Fallback: odafc (requires ODAFileConverter on PATH)
-        if _HAS_ODAFC:
-            return odafc.readfile(str(path))
-        raise RuntimeError(
-            "DWG import needs ODA File Converter. Set ODA_CONVERTER_EXE to the exe "
-            "or place dwg2dxf_wrapper.bat next to the script."
-        )
-    return ezdxf.readfile(str(path))  # DXF directly
-
-
-def dxf_to_structured(doc: Drawing) -> dict:
-    msp = doc.modelspace()
-    header = doc.header
-    insunits = header.get("$INSUNITS", 0)  # 1 = inches, 6 = meters, etc.
-    # Normalize a simple unit string for quoting
-    UNITS = {0:"unitless",1:"in",2:"ft",3:"mi",4:"mm",5:"cm",6:"m"}
-    units = UNITS.get(insunits, "unitless")
-
-    texts = []
-    for e in msp.query("TEXT"):
-        texts.append({"type":"TEXT","text":e.dxf.text, "xy":tuple(e.dxf.insert[:2])})
-    for e in msp.query("MTEXT"):
-        # .text is formatted; .plain_text() strips inline codes
-        texts.append({"type":"MTEXT","text":e.plain_text(), "xy":tuple(e.dxf.insert[:2])})
-    for e in msp.query("MLEADER"):
-        if e.context and e.context.mtext:
-            texts.append({"type":"MLEADER","text":e.context.mtext.plain_text()})
-
-    dims = []
-    for d in msp.query("DIMENSION"):
-        # d.dxf.text often contains the displayed text; fall back to raw if needed
-        label = (d.dxf.text or "").strip()
-        # Try to find diameter / radius / WxH patterns in the label
-        m = DIM_RE.search(label.replace(",", "."))
-        dims.append({"raw": label, "pattern": m.groups() if m else None})
-
-    # Quick & dirty material/thickness/title block scrapes from text blobs
-    material = next((t["text"].split(":")[-1].strip()
-                     for t in texts if re.search(r"\\bMATERIAL\\b", t["text"], re.I)), None)
-    thickness = next((re.search(r"(?:(?:THICK|THK)[\\s:=]*)([0-9.]+)\\s*(mm|in|inch|inches)?", t["text"], re.I)
-                      for t in texts), None)
-    if thickness:
-        thickness = {"value": float(thickness.group(1)), "units": (thickness.group(2) or units)}
-
-    return {
-        "source_units": units,
-        "texts": texts,
-        "dimensions": dims,
-        "material": material,
-        "thickness": thickness,
-    }
-
-def pdf_to_structured(pdf_path: Path, page_index: int = 0, dpi: int = 300):
-    doc = fitz.open(pdf_path)
-    page = doc[page_index]
-
-    # 1) TEXT blocks (vector text if present)
-    words = page.get_text("words")  # [(x0,y0,x1,y1,"word", block_no, line_no, word_no), ...]
-    text_blocks = page.get_text("blocks")  # paragraph-level
-    raw_text = page.get_text("text")       # simple text stream
-
-    # 2) Tables (handy for title blocks / BOMs)
-    tables = []
-    try:
-        tabs = page.find_tables()  # since PyMuPDF 1.23+
-        for t in tabs.tables:
-            tables.append(t.extract())
-    except Exception:
-        pass
-
-    # 3) Vector line-art (for figure geometry / bounding clusters)
-    drawings = page.get_drawings()  # vector paths, strokes, fills
-    # Optional: cluster into groups that form a single figure (1.24+)
-    # clusters = page.cluster_drawings()
-
-    # 4) Page raster for Qwen
-    zoom = dpi / 72.0
-    mat = fitz.Matrix(zoom, zoom)
-    pix = page.get_pixmap(matrix=mat, alpha=False)
-    png_path = pdf_path.with_suffix(f".p{page_index}.png")
-    pix.save(png_path)
-
-    return {
-        "raw_text": raw_text,
-        "blocks": text_blocks,
-        "tables": tables,
-        "drawings_count": len(drawings),
-        "image_path": str(png_path),
-    }
-
-QWEN_SYSTEM = """You are a quoting assistant. Only confirm or fill missing parameters.
-Prefer vector-extracted numbers over reading pixels. If units conflict, explain and ask for clarification.
-Return strict JSON with fields:
-{overall_size_mm, thickness_mm, material, hole_counts_by_dia_mm, slots, threads, tolerance_notes, finish, qty, uncertainties[]}"""
-
-def build_llm_payload(structured: dict, page_image_path: str | None):
-    messages = [
-        {"role":"system","content": QWEN_SYSTEM},
-        {"role":"user","content":[
-            {"type":"text","text":"Vector-extracted data (trusted):\n" + json.dumps(structured, ensure_ascii=False)},
-        ]}
-    ]
-    if page_image_path:
-        messages[1]["content"].append({"type":"image_url","image_url":{"url":"file://"+page_image_path}})
-    return messages
-
-
-
-# ==== OpenCascade compat (works with OCP OR OCC.Core) ====
-import tempfile, subprocess
-from pathlib import Path
-
-try:
-    # ---- OCP branch ----
-    from OCP.IFSelect import IFSelect_RetDone
-    from OCP.IGESControl import IGESControl_Reader
-    from OCP.ShapeFix import ShapeFix_Shape, ShapeFix_Solid
-    from OCP.BRepCheck import BRepCheck_Analyzer
-    from OCP.BRep import BRep_Tool        # OCP version
-    from OCP.BRepGProp import BRepGProp
-    from OCP.GProp import GProp_GProps
-    from OCP.Bnd import Bnd_Box
-    from OCP.BRepBndLib import BRepBndLib
-    from OCP.BRep import BRep_Builder
-    from OCP.BRepBuilderAPI import (
-        BRepBuilderAPI_MakeFace, BRepBuilderAPI_MakePolygon,
-        BRepBuilderAPI_Sewing, BRepBuilderAPI_MakeSolid,
-    )
-    from OCP.BRepPrimAPI import BRepPrimAPI_MakePrism
-    from OCP.BRepAlgoAPI import BRepAlgoAPI_Section
-    from OCP.BRepAdaptor import BRepAdaptor_Surface
-    
-    # ADD THESE TWO IMPORTS
-    from OCP.TopTools import TopTools_IndexedDataMapOfShapeListOfShape
-    from OCP.TopoDS import (
-        TopoDS_Shape, TopoDS_Edge, TopoDS_Face, TopoDS_Shell, TopoDS_Solid, TopoDS_Compound
-    )
-    from OCP.TopExp import TopExp_Explorer, TopExp
-    from OCP.TopAbs import TopAbs_FACE, TopAbs_EDGE, TopAbs_SOLID, TopAbs_SHELL, TopAbs_COMPOUND
-    from OCP.GeomAdaptor import GeomAdaptor_Surface
-    from OCP.GeomAbs import (
-        GeomAbs_Plane, GeomAbs_Cylinder, GeomAbs_Torus, GeomAbs_Cone,
-        GeomAbs_BSplineSurface, GeomAbs_BezierSurface
-    )
-    from OCP.ShapeAnalysis import ShapeAnalysis_Surface
-    from OCP.gp import gp_Pnt, gp_Vec, gp_Dir, gp_Pln
-    BACKEND_OCC = "OCP"
-
-    def BRepTools_UVBounds(face):
-        return BRepTools.UVBounds(face)
-
-
-    def _brep_read(path: str) -> TopoDS_Shape:
-        s = TopoDS_Shape()
-        builder = BRep_Builder()
-        if hasattr(BRepTools, "Read_s"):
-            ok = BRepTools.Read_s(s, str(path), builder)
-        else:
-            ok = BRepTools.Read(s, str(path), builder)
-        if ok is False:
-            raise RuntimeError("BREP read failed")
-        return s
-
-
-
-except Exception:
-    # ---- OCC.Core branch ----
-    from OCC.Core.STEPControl import STEPControl_Reader
-    from OCC.Core.IFSelect import IFSelect_RetDone
-    from OCC.Core.IGESControl import IGESControl_Reader
-    from OCC.Core.ShapeFix import ShapeFix_Shape, ShapeFix_Solid
-    from OCC.Core.BRepCheck import BRepCheck_Analyzer
-    from OCC.Core.BRep import BRep_Tool          # ? OCC version
-    from OCC.Core.GProp import GProp_GProps
-    from OCC.Core.Bnd import Bnd_Box
-    from OCC.Core.BRep import BRep_Builder
-    from OCC.Core.BRepBuilderAPI import (
-        BRepBuilderAPI_MakeFace, BRepBuilderAPI_MakePolygon,
-        BRepBuilderAPI_Sewing, BRepBuilderAPI_MakeSolid,
-    )
-    from OCC.Core.BRepPrimAPI import BRepPrimAPI_MakePrism
-    from OCC.Core.BRepAlgoAPI import BRepAlgoAPI_Section
-    from OCC.Core.BRepAdaptor import BRepAdaptor_Surface
-    # ADD TopTools import and TopoDS_Face for the fix below
-    from OCC.Core.TopTools import TopTools_IndexedDataMapOfShapeListOfShape
-    from OCC.Core.TopoDS import (
-        TopoDS_Shape, TopoDS_Edge, TopoDS_Face, TopoDS_Shell, TopoDS_Solid, TopoDS_Compound,
-        TopoDS_Face
-    )
-    from OCC.Core.TopExp import TopExp_Explorer
-    from OCC.Core.TopAbs import TopAbs_FACE, TopAbs_EDGE, TopAbs_SOLID, TopAbs_SHELL, TopAbs_COMPOUND
-    from OCC.Core.GeomAdaptor import GeomAdaptor_Surface
-    from OCC.Core.GeomAbs import (
-        GeomAbs_Plane, GeomAbs_Cylinder, GeomAbs_Torus, GeomAbs_Cone,
-        GeomAbs_BSplineSurface, GeomAbs_BezierSurface
-    )
-    from OCC.Core.ShapeAnalysis import ShapeAnalysis_Surface
-    from OCC.Core.gp import gp_Pnt, gp_Vec, gp_Dir, gp_Pln
-    BACKEND_OCC = "OCC.Core"
-
-    # BRepGProp shim (pythonocc uses free functions)
-    from OCC.Core.BRepGProp import (
-        brepgprop_SurfaceProperties,
-        brepgprop_LinearProperties,
-        brepgprop_VolumeProperties,
-    )
-    class _BRepGPropShim:
-        @staticmethod
-        def SurfaceProperties_s(shape_or_face, gprops): brepgprop_SurfaceProperties(shape_or_face, gprops)
-        @staticmethod
-        def LinearProperties_s(edge, gprops):          brepgprop_LinearProperties(edge, gprops)
-        @staticmethod
-        def VolumeProperties_s(shape, gprops):         brepgprop_VolumeProperties(shape, gprops)
-    BRepGProp = _BRepGPropShim
-
-
-
-    # UV bounds and brep read are free functions
-    from OCC.Core.BRepTools import BRepTools, breptools_Read as _occ_breptools_read
-    def BRepTools_UVBounds(face):
-        fn = getattr(BRepTools, "UVBounds", None)
-        if fn is None:
-            from OCC.Core.BRepTools import breptools_UVBounds as _legacy
-            return _legacy(face)
-        return fn(face)
-    def _brep_read(path: str) -> TopoDS_Shape:
-        s = TopoDS_Shape()
-        ok = _occ_breptools_read(s, str(path), BRep_Builder())
-        if ok is False:
-            raise RuntimeError("BREP read failed")
-        return s
-
-def _shape_from_reader(reader):
-    """Return a healed TopoDS_Shape from a STEP/IGES reader."""
-    transfer_count = 0
-    if hasattr(reader, "NbShapes"):
-        try:
-            transfer_count = reader.NbShapes()
-        except Exception:
-            transfer_count = 0
-    if not transfer_count and hasattr(reader, "NbRootsForTransfer"):
-        try:
-            transfer_count = reader.NbRootsForTransfer()
-        except Exception:
-            transfer_count = 0
-    if transfer_count <= 0:
-        raise RuntimeError("Reader produced zero shapes")
-
-    if transfer_count == 1:
-        shape = reader.Shape(1)
-    else:
-        builder = BRep_Builder()
-        compound = TopoDS_Compound()
-        builder.MakeCompound(compound)
-        added = 0
-        for i in range(1, transfer_count + 1):
-            s = reader.Shape(i)
-            if s is None or s.IsNull():
-                continue
-            builder.Add(compound, s)
-            added += 1
-        if added == 0:
-            raise RuntimeError("Reader produced only null sub-shapes")
-        shape = compound
-
-    if shape is None or shape.IsNull():
-        raise RuntimeError("Reader produced a null TopoDS_Shape")
-
-    fixer = ShapeFix_Shape(shape)
-    fixer.Perform()
-    healed = fixer.Shape()
-    if healed is None or healed.IsNull():
-        raise RuntimeError("Shape healing failed (null shape)")
-
-    try:
-        analyzer = BRepCheck_Analyzer(healed)
-        # we do not require validity, but invoking the analyzer surfaces issues early
-        analyzer.IsValid()
-    except Exception:
-        pass
-
-    return healed
-
-def read_step_shape(path: str) -> TopoDS_Shape:
-    rdr = STEPControl_Reader()
-    if rdr.ReadFile(path) != IFSelect_RetDone:
-        raise RuntimeError("STEP read failed (file unreadable or unsupported).")
-    rdr.TransferRoots()
-    n = rdr.NbShapes()
-    if n == 0:
-        raise RuntimeError("STEP read produced zero shapes (no transferable roots).")
-
-    if n == 1:
-        shape = rdr.Shape(1)
-    else:
-        builder = BRep_Builder()
-        comp = TopoDS_Compound()
-        builder.MakeCompound(comp)
-        for i in range(1, n + 1):
-            s = rdr.Shape(i)
-            if not s.IsNull():
-                builder.Add(comp, s)
-        shape = comp
-
-    if shape.IsNull():
-        raise RuntimeError("STEP produced a null TopoDS_Shape.")
-    # Verify we truly pass a Shape to MapShapesAndAncestors
-    print("[DBG] shape type:", type(shape).__name__, "IsNull:", getattr(shape, "IsNull", lambda: True)())
-    amap = map_shapes_and_ancestors(shape, TopAbs_EDGE, TopAbs_FACE)
-    print("[DBG] map size:", amap.Size())
-    # DEBUG: sanity probe for STEP faces
-    if os.environ.get("STEP_PROBE", "0") == "1":
-        cnt = 0
-        try:
-            for f in iter_faces(shape):
-                _surf, _loc = face_surface(f)
-                cnt += 1
-        except Exception as _e:
-            # Keep debug non-fatal; report and continue
-            print(f"[STEP_PROBE] error during face probe: {_e}")
-        else:
-            print(f"[STEP_PROBE] faces={cnt}")
-
-    fx = ShapeFix_Shape(shape)
-    fx.Perform()
-    return fx.Shape()
-
-from OCP.TopoDS import TopoDS_Shape  # or OCC.Core.TopoDS on pythonocc
-
-def safe_bbox(shape: TopoDS_Shape):
-    if shape is None or shape.IsNull():
-        raise ValueError("Cannot compute bounding box of a null shape.")
-    box = Bnd_Box()
-    bnd_add(shape, box, True)  # <- uses whichever binding is available
-    return box
-def read_step_or_iges_or_brep(path: str) -> TopoDS_Shape:
-    p = Path(path)
-    ext = p.suffix.lower()
-    if ext in (".step", ".stp"):
-        return read_step_shape(str(p))
-    if ext in (".iges", ".igs"):
-        ig = IGESControl_Reader()
-        if ig.ReadFile(str(p)) != IFSelect_RetDone:
-            raise RuntimeError("IGES read failed")
-        ig.TransferRoots()
-        return _shape_from_reader(ig)
-    if ext == ".brep":
-        return _brep_read(str(p))
-    raise RuntimeError(f"Unsupported OCC format: {ext}")
-
-def convert_dwg_to_dxf(dwg_path: str, *, out_ver="ACAD2018") -> str:
-    """
-    Robust DWG?DXF wrapper.
-    Works with:
-      - A .bat/.cmd wrapper that accepts:  <input.dwg> <output.dxf>
-      - A custom exe that accepts:         <input.dwg> <output.dxf>
-      - ODAFileConverter.exe (7-arg form): <in_dir> <out_dir> <ver> DXF 0 0 <filter>
-    Looks for the converter via env var or common local paths and prints
-    the exact command used on failure.
-    """
-    # 1) find converter
-    exe = (os.environ.get("ODA_CONVERTER_EXE")
-           or os.environ.get("DWG2DXF_EXE")
-           or str(Path(__file__).with_name("dwg2dxf_wrapper.bat"))
-           or r"D:\CAD_Quoting_Tool\dwg2dxf_wrapper.bat")
-
-    if not exe or not Path(exe).exists():
-        raise RuntimeError(
-            "DWG import needs a DWG?DXF converter.\n"
-            "Set ODA_CONVERTER_EXE (recommended) or DWG2DXF_EXE to a .bat/.cmd/.exe.\n"
-            "Expected .bat signature:  <input.dwg> <output.dxf>"
-        )
-
-    dwg = Path(dwg_path)
-    out_dir = Path(tempfile.mkdtemp(prefix="dwg2dxf_"))
-    out_dxf = out_dir / (dwg.stem + ".dxf")
-
-    exe_lower = Path(exe).name.lower()
-    try:
-        if exe_lower.endswith(".bat") or exe_lower.endswith(".cmd"):
-            # ? run batch via cmd.exe so it actually executes
-            cmd = ["cmd", "/c", exe, str(dwg), str(out_dxf)]
-            proc = subprocess.run(cmd, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
-        elif "odafileconverter" in exe_lower:
-            # ? official ODAFileConverter CLI
-            cmd = [exe, str(dwg.parent), str(out_dir), out_ver, "DXF", "0", "0", dwg.name]
-            proc = subprocess.run(cmd, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
-        else:
-            # ? generic exe that accepts <in> <out>
-            cmd = [exe, str(dwg), str(out_dxf)]
-            proc = subprocess.run(cmd, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
-    except subprocess.CalledProcessError as e:
-        raise RuntimeError(
-            "DWG?DXF conversion failed.\n"
-            f"cmd: {' '.join(cmd)}\n"
-            f"stdout:\n{e.stdout}\n"
-            f"stderr:\n{e.stderr}"
-        ) from e
-
-    # 2) resolve the produced DXF
-    produced = out_dxf if out_dxf.exists() else (out_dir / (dwg.stem + ".dxf"))
-    if not produced.exists():
-        raise RuntimeError(
-            "Converter returned success but DXF not found.\n"
-            f"cmd: {' '.join(cmd)}\n"
-            f"checked: {out_dxf} | {produced}"
-        )
-    return str(produced)
-
-
-
-ANG_TOL = math.radians(5.0)
-DOT_TOL = math.cos(ANG_TOL)
-SMALL = 1e-7
-
-
-
-def iter_solids(shape: TopoDS_Shape):
-    exp = TopExp_Explorer(shape, TopAbs_SOLID)
-    while exp.More():
-        yield to_solid(exp.Current())
-        exp.Next()
-
-def explode_compound(shape: TopoDS_Shape):
-    """If the file is a big COMPOUND, break it into shapes (parts/bodies)."""
-    exp = TopExp_Explorer(shape, TopAbs_COMPOUND)
-    if exp.More():
-        # Itï¿½s a compound ï¿½ return its shells/solids/faces as needed
-        solids = list(iter_solids(shape))
-        if solids:
-            return solids
-        # fallback to shells
-        sh = TopExp_Explorer(shape, TopAbs_SHELL)
-        shells = []
-        while sh.More():
-            shells.append(to_shell(sh.Current()))
-            sh.Next()
-        return shells
-    return [shape]
-
-
-
-def safe_bounding_box(shape: TopoDS_Shape) -> Bnd_Box:
-    """Public wrapper maintained for compatibility."""
-    return safe_bbox(shape)
-
-def _bbox(shape):
-    box = safe_bbox(shape)
-    xmin, ymin, zmin, xmax, ymax, zmax = box.Get()
-    return (xmin, ymin, zmin, xmax, ymax, zmax)
-
-def _area_of_face(face) -> float:
-    props = GProp_GProps()
-    BRepGProp.SurfaceProperties_s(face, props)
-    return float(props.Mass())
-
-def _length_of_edge(edge) -> float:
-    props = GProp_GProps()
-    linear_properties(edge, props)
-    return float(props.Mass())
-
-def _face_midpoint_uv(face):
-    umin, umax, vmin, vmax = BRepTools_UVBounds(face)
-    return (0.5*(umin+umax), 0.5*(vmin+vmax))
-
-
-def _face_normal(face):
-    try:
-        try:
-            from OCP.BRepLProp import BRepLProp_SLProps
-        except ImportError:
-            from OCC.Core.BRepLProp import BRepLProp_SLProps
-        u, v = _face_midpoint_uv(face)
-        props = BRepLProp_SLProps(face_surface(face)[0], u, v, 1, SMALL)
-        if props.IsNormalDefined():
-            n = props.Normal()
-            if face.Orientation().Name() == "REVERSED":
-                n.Reverse()
-            return gp_Dir(n.X(), n.Y(), n.Z())
-    except Exception:
-        pass
-    return None
-
-def _face_type(face) -> str:
-    surf = face_surface(face)[0]
-    ga = GeomAdaptor_Surface(surf)
-    st = ga.GetType()
-    if st == GeomAbs_Plane: return "planar"
-    if st in (GeomAbs_Cylinder, GeomAbs_Torus, GeomAbs_Cone): return "cylindrical"
-    if st in (GeomAbs_BSplineSurface, GeomAbs_BezierSurface): return "freeform"
-    return "other"
-
-
-
-def _cluster_normals(normals):
-    clusters = []
-    for n in normals:
-        added = False
-        for c in clusters:
-            if abs(n.Dot(c)) > DOT_TOL:
-                added = True; break
-        if not added:
-            clusters.append(n)
-    return clusters
-
-def _sum_edge_length_sharp(shape, angle_thresh_deg=175.0) -> float:
-    angle_thresh = math.radians(angle_thresh_deg)
-    edge2faces = map_shapes_and_ancestors(shape, TopAbs_EDGE, TopAbs_FACE)
-    total = 0.0
-    for i in range(1, map_size(edge2faces) + 1):
-        edge = to_edge(edge2faces.FindKey(i))
-        face_list = edge2faces.FindFromIndex(i)
-        faces = [ensure_face(shp) for shp in list_iter(face_list)]
-        if len(faces) < 2:
-            continue
-        f1, f2 = faces[0], faces[-1]
-        n1 = _face_normal(f1)
-        n2 = _face_normal(f2)
-        if not n1 or not n2:
-            continue
-        ang = math.acos(max(-1.0, min(1.0, abs(n1.Dot(n2)))))
-        if ang < (math.pi - angle_thresh):
-            total += _length_of_edge(edge)
-    return total
-
-def _largest_planar_faces_and_normals(shape):
-    largest_area = 0.0
-    normals = []
-    for f in iter_faces(shape):
-        if _face_type(f) == "planar":
-            a = _area_of_face(f)
-            largest_area = max(largest_area, a)
-            n = _face_normal(f)
-            if n: normals.append(n)
-    clusters = _cluster_normals(normals)
-    return largest_area, clusters
-
-def _surface_areas_by_type(shape):
-    from collections import defaultdict
-    areas = defaultdict(float)
-    for f in iter_faces(shape):
-        t = _face_type(f)
-        areas[t] += _area_of_face(f)
-    return areas
-
-def _section_perimeter_len(shape, z_values):
-    xmin, ymin, zmin, xmax, ymax, zmax = _bbox(shape)
-    total = 0.0
-    for z in z_values:
-        plane = gp_Pln(gp_Pnt(0,0,z), gp_Dir(0,0,1))
-        sec = BRepAlgoAPI_Section(shape, plane, False); sec.Build()
-        if not sec.IsDone(): continue
-        w = sec.Shape()
-        it = TopExp_Explorer(w, TopAbs_EDGE)
-        while it.More():
-            e = to_edge(it.Current())
-            total += _length_of_edge(e)
-            it.Next()
-    return total
-
-def _min_wall_between_parallel_planes(shape):
-    planes = []
-    for f in iter_faces(shape):
-        if _face_type(f) == "planar":
-            n = _face_normal(f)
-            if n:
-                umin, umax, vmin, vmax = BRepTools_UVBounds(f)
-                surf, _ = face_surface(f)
-                sas = ShapeAnalysis_Surface(surf)
-                pnt = sas.Value(0.5*(umin+umax), 0.5*(vmin+vmax))
-                d = n.X()*pnt.X() + n.Y()*pnt.Y() + n.Z()*pnt.Z()
-                planes.append((f, n, d, _bbox(f)))
-    def overlap(a1, a2, b1, b2): return not (a2 < b1 or b2 < a1)
-    min_th = None
-    for i in range(len(planes)):
-        f1, n1, d1, b1 = planes[i]
-        for j in range(i+1, len(planes)):
-            f2, n2, d2, b2 = planes[j]
-            if abs(abs(n1.Dot(n2)) - 1.0) < 1e-3:
-                nd = (abs(n1.X()), abs(n1.Y()), abs(n1.Z()))
-                normal_axis = nd.index(max(nd))
-                if normal_axis == 2:
-                    ok = overlap(b1[0], b1[3], b2[0], b2[3]) and overlap(b1[1], b1[4], b2[1], b2[4])
-                elif normal_axis == 1:
-                    ok = overlap(b1[0], b1[3], b2[0], b2[3]) and overlap(b1[2], b1[5], b2[2], b2[5])
-                else:
-                    ok = overlap(b1[1], b1[4], b2[1], b2[4]) and overlap(b1[2], b1[5], b2[2], b2[5])
-                if ok:
-                    th = abs(d1 - d2)
-                    if (min_th is None) or (th < min_th): min_th = th
-    return min_th
-
-def _hole_groups_from_cylinders(shape, bbox=None):
-    if bbox is None: bbox = _bbox(shape)
-    from collections import defaultdict
-    groups = defaultdict(lambda: {"dia_mm":0.0,"depth_mm":0.0,"through":False,"count":0})
-    for f in iter_faces(shape):
-        if _face_type(f) == "cylindrical":
-            ga = GeomAdaptor_Surface(face_surface(f)[0])
-            try:
-                cyl = ga.Cylinder(); r = abs(cyl.Radius()); ax = cyl.Axis().Direction()
-                fb = _bbox(f)
-                def proj(x,y,z): return x*ax.X() + y*ax.Y() + z*ax.Z()
-                span = abs(proj(fb[3],fb[4],fb[5]) - proj(fb[0],fb[1],fb[2]))
-                dia = 2.0*r
-                bmin = proj(*bbox[:3]); bmax = proj(*bbox[3:]); bspan = abs(bmax-bmin)
-                through = span > 0.9*bspan
-                key = (round(dia,2), round(span,2), through)
-                groups[key]["dia_mm"] = round(dia,2)
-                groups[key]["depth_mm"] = round(span,2)
-                groups[key]["through"] = through
-                groups[key]["count"] += 1
-            except Exception:
-                pass
-    return list(groups.values())
-
-def _turning_score(shape, areas_by_type):
-    total = sum(areas_by_type.values()) or 1.0
-    cyl_ratio = areas_by_type.get("cylindrical", 0.0)/total
-    axes = []
-    for f in iter_faces(shape):
-        if _face_type(f) == "cylindrical":
-            ga = GeomAdaptor_Surface(face_surface(f)[0])
-            try: axes.append(ga.Cylinder().Axis().Direction())
-            except Exception: pass
-    if axes:
-        ax = gp_Dir(sum(a.X() for a in axes) or 1e-9, sum(a.Y() for a in axes) or 1e-9, sum(a.Z() for a in axes) or 1e-9)
-        align = sum(abs(ax.Dot(a)) for a in axes)/len(axes)
-    else:
-        align = 0.0
-    score = max(0.0, min(1.0, 0.5*cyl_ratio + 0.5*align))
-    xmin,ymin,zmin,xmax,ymax,zmax = _bbox(shape)
-    Lx, Ly, Lz = xmax-xmin, ymax-ymin, zmax-zmin
-    length = max(Lx, Ly, Lz)
-    maxod = sorted([Lx, Ly, Lz])[-2]
-    return {"GEO_Turning_Score_0to1": round(score,3), "GEO_MaxOD_mm": round(maxod,3), "GEO_Length_mm": round(length,3)}
-
-def enrich_geo_occ(shape):
-    xmin,ymin,zmin,xmax,ymax,zmax = _bbox(shape)
-    L,W,H = xmax-xmin, ymax-ymin, zmax-zmin
-
-    props = GProp_GProps()
-    try:
-        BRepGProp.VolumeProperties_s(shape, props); volume = float(props.Mass())
-        com = props.CentreOfMass(); center = [com.X(), com.Y(), com.Z()]
-    except Exception:
-        volume, center = 0.0, [0.0,0.0,0.0]
-    try:
-        BRepGProp.SurfaceProperties_s(shape, props); area_total = float(props.Mass())
-    except Exception:
-        area_total = 0.0
-
-    faces = 0
-    for _ in iter_faces(shape): faces += 1
-
-    areas_by_type = _surface_areas_by_type(shape)
-    largest_planar, normal_clusters = _largest_planar_faces_and_normals(shape)
-    unique_normals = len(normal_clusters)
-
-    min_wall = _min_wall_between_parallel_planes(shape)
-    thinwall_present = (min_wall is not None) and (min_wall < 1.0)
-
-    z_vals = [zmin + 0.25*(zmax-zmin), zmin + 0.50*(zmax-zmin), zmin + 0.75*(zmax-zmin)]
-    wedm_len = _section_perimeter_len(shape, z_vals)
-
-    deburr_len = _sum_edge_length_sharp(shape, angle_thresh_deg=175.0)
-
-    holes = _hole_groups_from_cylinders(shape, (xmin,ymin,zmin,xmax,ymax,zmax))
-
-    bbox_vol = max(L*W*H, 1e-9)
-    free_ratio = areas_by_type.get("freeform", 0.0)/max(area_total, 1e-9)
-    complexity = (faces / (volume if volume>0 else bbox_vol))*100.0 + 30.0*free_ratio
-
-    axis_dirs = [gp_Dir(1,0,0), gp_Dir(-1,0,0), gp_Dir(0,1,0), gp_Dir(0,-1,0), gp_Dir(0,0,1), gp_Dir(0,0,-1)]
-    hit=0; tot=0
-    for f in iter_faces(shape):
-        n = _face_normal(f)
-        if n:
-            tot += 1
-            if any(abs(n.Dot(a))>DOT_TOL for a in axis_dirs): hit += 1
-    access = (hit/tot) if tot else 0.0
-
-    geo = {
-        "GEO-01_Length_mm": round(L,3),
-        "GEO-02_Width_mm": round(W,3),
-        "GEO-03_Height_mm": round(H,3),
-        "GEO-Volume_mm3": round(volume,2),
-        "GEO-SurfaceArea_mm2": round(area_total,2),
-        "Feature_Face_Count": int(faces),
-        "GEO_Area_Planar_mm2": round(areas_by_type.get("planar",0.0),2),
-        "GEO_Area_Cyl_mm2": round(areas_by_type.get("cylindrical",0.0),2),
-        "GEO_Area_Freeform_mm2": round(areas_by_type.get("freeform",0.0),2),
-        "GEO_LargestPlane_Area_mm2": round(largest_planar,2),
-        "GEO_Setup_UniqueNormals": int(unique_normals),
-        "GEO_MinWall_mm": round(min_wall,3) if min_wall is not None else None,
-        "GEO_ThinWall_Present": bool(thinwall_present),
-        "GEO_WEDM_PathLen_mm": round(wedm_len,2),
-        "GEO_Deburr_EdgeLen_mm": round(deburr_len,2),
-        "GEO_Hole_Groups": holes,
-        "GEO_Complexity_0to100": round(complexity,2),
-        "GEO_3Axis_Accessible_Pct": round(access,3),
-        "GEO_CenterOfMass": center,
-        "OCC_Backend": "OCC"
-    }
-    geo.update(_turning_score(shape, areas_by_type))
-    return geo
-
-def enrich_geo_stl(path):
-    import time
-    start_time = time.time()
-    print(f"[{time.time() - start_time:.2f}s] Starting enrich_geo_stl for {path}")
-    if not _HAS_TRIMESH:
-        raise RuntimeError("trimesh not available to process STL")
-    
-    print(f"[{time.time() - start_time:.2f}s] Loading mesh...")
-    mesh = trimesh.load(path, force='mesh')
-    print(f"[{time.time() - start_time:.2f}s] Mesh loaded. Faces: {len(mesh.faces)}")
-    
-    if mesh.is_empty:
-        raise RuntimeError("Empty STL mesh")
-    
-    (xmin, ymin, zmin), (xmax, ymax, zmax) = mesh.bounds
-    L, W, H = float(xmax-xmin), float(ymax-ymin), float(zmax-zmin)
-    area_total = float(mesh.area)
-    volume = float(mesh.volume) if mesh.is_volume else 0.0
-    faces = int(len(mesh.faces))
-    
-    print(f"[{time.time() - start_time:.2f}s] Calculating WEDM length...")
-    z_vals = [zmin + 0.25*(zmax-zmin), zmin + 0.50*(zmax-zmin), zmin + 0.75*(zmax-zmin)]
-    wedm_len = 0.0
-    for z in z_vals:
-        sec = mesh.section(plane_origin=[0,0,z], plane_normal=[0,0,1])
-        if sec is None:
-            continue
-        planar = sec.to_2D() # Changed to_planar() to to_2D()
-        try:
-            wedm_len += float(planar.length)
-        except Exception:
-            pass
-    print(f"[{time.time() - start_time:.2f}s] WEDM length calculated.")
-    
-    print(f"[{time.time() - start_time:.2f}s] Calculating deburr length...")
-    try:
-        import numpy as np
-        angles = mesh.face_adjacency_angles
-        edges = mesh.face_adjacency_edges
-        if len(angles) and len(edges):
-            sharp = angles > math.radians(15.0)
-            vec = mesh.vertices[edges[:,0]] - mesh.vertices[edges[:,1]]
-            lengths = np.linalg.norm(vec, axis=1)
-            deburr_len = float(lengths[sharp].sum())
-        else:
-            deburr_len = 0.0
-    except Exception:
-        deburr_len = 0.0
-    print(f"[{time.time() - start_time:.2f}s] Deburr length calculated.")
-    
-    bbox_vol = max(L*W*H, 1e-9)
-    complexity = (faces / max(volume, bbox_vol)) * 100.0
-    
-    print(f"[{time.time() - start_time:.2f}s] Calculating 3-axis accessibility...")
-    try:
-        n = mesh.face_normals
-        area_faces = mesh.area_faces.reshape(-1,1)
-        import numpy as np
-        axes = np.array([[1,0,0],[-1,0,0],[0,1,0],[0,-1,0],[0,0,1],[0,0,-1]], dtype=float).T
-        dots = np.abs(n @ axes)  # (F,6)
-        close = (dots > DOT_TOL).any(axis=1)
-        access = float((area_faces[close].sum() / area_faces.sum())) if area_faces.sum() > 0 else 0.0
-    except Exception:
-        access = 0.0
-    print(f"[{time.time() - start_time:.2f}s] 3-axis accessibility calculated.")
-    
-    try:
-        center = list(map(float, mesh.center_mass))
-    except Exception:
-        center = [0.0,0.0,0.0]
-    
-    print(f"[{time.time() - start_time:.2f}s] Finished enrich_geo_stl.")
-    return {
-        "GEO-01_Length_mm": round(L,3),
-        "GEO-02_Width_mm": round(W,3),
-        "GEO-03_Height_mm": round(H,3),
-        "GEO-Volume_mm3": round(volume,2),
-        "GEO-SurfaceArea_mm2": round(area_total,2),
-        "Feature_Face_Count": faces,
-        "GEO_LargestPlane_Area_mm2": None,
-        "GEO_Setup_UniqueNormals": None,
-        "GEO_MinWall_mm": None,
-        "GEO_ThinWall_Present": False,
-        "GEO_WEDM_PathLen_mm": round(wedm_len,2),
-        "GEO_Deburr_EdgeLen_mm": round(deburr_len,2),
-        "GEO_Hole_Groups": [],
-        "GEO_Complexity_0to100": round(complexity,2),
-        "GEO_3Axis_Accessible_Pct": round(access,3),
-        "GEO_CenterOfMass": center,
-        "OCC_Backend": "trimesh (STL)"
-    }
-
-
-def load_cad_any(path: str) -> TopoDS_Shape:
-    ext = Path(path).suffix.lower().lstrip(".")
-    if ext in ("step", "stp", "iges", "igs", "brep"):
-        return read_step_or_iges_or_brep(path)
-    if ext == "dwg":
-        dxf_path = convert_dwg_to_dxf(path)
-        return read_dxf_as_occ_shape(dxf_path)
-    if ext == "dxf":
-        return read_dxf_as_occ_shape(path)
-    
-    raise RuntimeError(f"Unsupported file type for shape loading: {ext}")
-def read_cad_any(path: str):
-    from OCP.IFSelect import IFSelect_RetDone
-    from OCP.ShapeFix import ShapeFix_Shape
-    from OCP.IGESControl import IGESControl_Reader
-    from OCP.TopoDS import TopoDS_Shape
-
-    ext = Path(path).suffix.lower()
-    if ext in (".step", ".stp"):
-        return read_step_shape(path)
-    if ext in (".iges", ".igs"):
-        ig = IGESControl_Reader()
-        if ig.ReadFile(path) != IFSelect_RetDone:
-            raise RuntimeError("IGES read failed")
-        ig.TransferRoots()
-        return _shape_from_reader(ig)
-    if ext == ".brep":
-        s = TopoDS_Shape()
-        if not BRepTools.Read(s, path, None):
-            raise RuntimeError("BREP read failed")
-        return s
-    if ext == ".dxf":
-        return read_dxf_as_occ_shape(path)
-    if ext == ".dwg":
-        conv = os.environ.get("ODA_CONVERTER_EXE") or os.environ.get("DWG2DXF_EXE")
-        print(f"INFO: Using DWG converter: {conv}")
-        dxf_path = convert_dwg_to_dxf(path)
-        return read_dxf_as_occ_shape(dxf_path)
-    raise RuntimeError(f"Unsupported CAD format: {ext}")
 
 # ----------------- Offline Qwen via llama-cpp -----------------
 def _llama_missing_msg():
@@ -7407,106 +6065,6 @@ def edit_variables_tk(df):
         messagebox.showinfo("Saved","Variables updated in memory; Save in app to persist.")
     tree.bind("<Button-1>", on_click); entry.bind("<Return>", on_return); ttk.Button(win,text="Save",command=on_save).pack()
     win.grab_set(); win.wait_window(); return df
-def read_dxf_as_occ_shape(dxf_path: str):
-    # minimal DXF?OCC triangulated shape (3DFACE/MESH/POLYFACE), fallback: extrude closed polyline
-    import ezdxf, numpy as np
-    from OCP.BRepBuilderAPI import (BRepBuilderAPI_MakeFace, BRepBuilderAPI_MakePolygon,
-                                    BRepBuilderAPI_Sewing, BRepBuilderAPI_MakeSolid)
-    from OCP.gp import gp_Pnt, gp_Vec
-    from OCP.BRepPrimAPI import BRepPrimAPI_MakePrism
-    from OCP.ShapeFix import ShapeFix_Solid
-    from OCP.TopExp import TopExp_Explorer
-    from OCP.TopAbs import TopAbs_FACE
-
-    def tri_face(p0,p1,p2):
-        poly = BRepBuilderAPI_MakePolygon()
-        for p in (p0,p1,p2,p0):
-            poly.Add(gp_Pnt(*p))
-        return BRepBuilderAPI_MakeFace(poly.Wire(), True).Face()
-
-    def sew(faces):
-        sew = BRepBuilderAPI_Sewing(1.0e-6, True, True, True, True)
-        for f in faces: sew.Add(f)
-        sew.Perform()
-        return sew.SewedShape()
-
-    doc = ezdxf.readfile(dxf_path)
-    msp = doc.modelspace()
-    INSUNITS = doc.header.get("$INSUNITS", 1)  # 1=in, 4=mm, 2=ft, 6=m
-    u2mm = {1:25.4, 4:1.0, 2:304.8, 6:1000.0}.get(INSUNITS, 1.0)
-
-    tris = []
-    # 3DFACE
-    for e in msp.query("3DFACE"):
-        verts = [e.dxf.vtx0, e.dxf.vtx1, e.dxf.vtx2, e.dxf.vtx3]
-        pts = [(vx * u2mm, vy * u2mm, vz * u2mm) for (vx, vy, vz) in verts]
-        if len(pts) < 3:
-            continue
-        tris.append((pts[0], pts[1], pts[2]))
-        if len(pts) == 4:
-            v3 = np.array(pts[2])
-            v4 = np.array(pts[3])
-            if np.linalg.norm(v3 - v4) > 1e-12:
-                tris.append((pts[0], pts[2], pts[3]))
-    # POLYFACE
-    for e in msp.query("POLYFACE"):
-        for f in e.faces():
-            pts=[(v.dxf.location.x*u2mm, v.dxf.location.y*u2mm, v.dxf.location.z*u2mm) for v in f.vertices()]
-            if len(pts)>=3:
-                tris.append((pts[0], pts[1], pts[2]))
-                for k in range(3,len(pts)):
-                    tris.append((pts[0], pts[k-1], pts[k]))
-    # MESH
-    for e in msp.query("MESH"):
-        for f in e.faces_as_vertices():
-            pts=[(v.x*u2mm, v.y*u2mm, v.z*u2mm) for v in f]
-            if len(pts)>=3:
-                tris.append((pts[0], pts[1], pts[2]))
-                for k in range(3,len(pts)):
-                    tris.append((pts[0], pts[k-1], pts[k]))
-
-    faces = [tri_face(*t) for t in tris]
-    shape = None
-    if faces:
-        sewed = sew(faces)
-        try:
-            solid = BRepBuilderAPI_MakeSolid()
-            exp = TopExp_Explorer(sewed, TopAbs_FACE)
-            while exp.More():
-                solid.Add(exp.Current()); exp.Next()
-            fix = ShapeFix_Solid(solid.Solid()); fix.Perform()
-            shape = fix.Solid()
-        except Exception:
-            shape = sewed
-
-    # Fallback: 2D closed polyline ? extrude small thickness
-    if (shape is None) or shape.IsNull():
-        for pl in msp.query("LWPOLYLINE"):
-            if pl.closed:
-                pts2d = [(x*u2mm, y*u2mm, 0.0) for x,y,_ in pl.get_points("xyb")]
-                poly = BRepBuilderAPI_MakePolygon()
-                for q in pts2d: poly.Add(gp_Pnt(*q))
-                poly.Close()
-                face = BRepBuilderAPI_MakeFace(poly.Wire(), True).Face()
-                thk_mm = float(os.environ.get("DXF_EXTRUDE_THK_MM", "5.0"))
-                shape = BRepPrimAPI_MakePrism(face, gp_Vec(0,0,thk_mm)).Shape()
-                break
-
-    if (shape is None) or shape.IsNull():
-        raise RuntimeError("DXF contained no 3D geometry I can use. Prefer STEP/SAT if possible.")
-    return shape
-
-# ---- 2D: PDF (PyMuPDF) -------------------------------------------------------
-try:
-    import fitz  # old import name
-    _HAS_PYMUPDF = True
-except Exception:
-    try:
-        import pymupdf as fitz  # new import name
-        _HAS_PYMUPDF = True
-    except Exception:
-        fitz = None  # allow the rest of the app to import
-
 def extract_2d_features_from_pdf_vector(pdf_path: str) -> dict:
     if not _HAS_PYMUPDF:
         raise RuntimeError("PyMuPDF (fitz) not installed. pip install pymupdf")
@@ -11289,10 +9847,12 @@ class ScrollableFrame(ttk.Frame):
 
 
 class App(tk.Tk):
-    def __init__(self):
+    def __init__(self, geometry_service: Optional[geometry.GeometryService] = None):
         super().__init__()
         self.title("Compos-AI")
         self.geometry("1260x900")
+
+        self.geometry_service = geometry_service or geometry.GeometryService()
 
         self.vars_df = None
         self.vars_df_full = None
@@ -12002,7 +10562,7 @@ class App(tk.Tk):
         geo = None
         try:
             # Fast path (your OCC feature extractor for 3D)
-            geo = extract_features_with_occ(path)  # handles STEP/IGES/BREP
+            geo = self.geometry_service.extract_occ_features(path)  # handles STEP/IGES/BREP
         except Exception:
             geo = None
 
@@ -12010,7 +10570,7 @@ class App(tk.Tk):
             if ext == ".stl":
                 stl_geo = None
                 try:
-                    stl_geo = enrich_geo_stl(path)  # trimesh-based
+                    stl_geo = self.geometry_service.enrich_stl(path)  # trimesh-based
                 except Exception as e:
                     import traceback
                     tb = traceback.format_exc()
@@ -12031,11 +10591,11 @@ class App(tk.Tk):
             else:
                 try:
                     if ext in (".step", ".stp"):
-                        shape = read_step_shape(path)
+                        shape = self.geometry_service.read_step(path)
                     else:
-                        shape = read_cad_any(path)            # IGES/BREP and others
+                        shape = self.geometry_service.read_model(path)            # IGES/BREP and others
                     _ = safe_bbox(shape)
-                    g = enrich_geo_occ(shape)             # OCC-based geometry features
+                    g = self.geometry_service.enrich_occ(shape)             # OCC-based geometry features
                     geo = _map_geo_to_double_underscore(g)
                 except Exception as e:
                     messagebox.showerror(
@@ -12469,7 +11029,8 @@ def _main(argv: Optional[Sequence[str]] = None) -> int:
     if args.no_gui:
         return 0
 
-    App().mainloop()
+    geo_service = geometry.GeometryService()
+    App(geometry_service=geo_service).mainloop()
     return 0
