+++ cad_quoter/__init__.py
@@ -1,0 +1,5 @@
+"""CAD Quoter package root."""
+
+from . import geometry  # re-export for convenience
+
+__all__ = ["geometry"]
